requirements.txt
fastapi==0.115.12
httpx==0.28.1
langchain==0.3.23
numpy==2.2.4
openai==1.74.0
pandas==2.2.3
promptflow==1.17.2
promptflow_tools==1.6.0
prompty==0.1.50
pydantic==2.11.3
pytest==8.3.5
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
sentence_transformers==4.0.2
watchdog==4.0.0
uvicorn==0.34.0
gunicorn==21.2.0
